+++ WebColliersCore/Models/pagosluz.cs
@@ -97,7 +97,7 @@ namespace WebLomelinCore.Models
 
         public B_inmuebles InmuebleData {  get; set; }
         public pagosluz ConsumoAnterior { get; set; }
-        public double? PorcentajeConsumo { get; set; }
+        public decimal? PorcentajeConsumo { get; set; }
         public string ClassPorcentaje { get; set; }
 
         /*DATOS DEL FORMULARIO*/
