+++ WebColliersCore/Models/pagospredial.cs
@@ -119,7 +119,7 @@ namespace WebLomelinCore.Models
 
         public B_inmuebles InmuebleData { get; set; }
         public pagospredial ConsumoAnterior { get; set; }
-        public double? PorcentajeConsumo { get; set; }
+        public decimal? PorcentajeConsumo { get; set; }
         public string ClassPorcentaje { get; set; }
